requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
qdrant-client==1.8.0
openai==1.14.0
twilio==9.0.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
